feat(blog): add draft post on numbered headings
This commit introduces a new blog post titled "Numbered Headings," which discusses a Chrome extension that adds hierarchical numbers to HTML heading tags. The post is currently in draft status and includes placeholders for media content such as YouTube videos and animated GIFs. A link to the source code repository is also provided.